requirements.txt
Django
django-cors-headers
djangorestframework
google-api-python-client
python-dateutil
